refactor(frontend): split chart and dataset table hooks
