+++ crates/fricon-ui/frontend/src/features/chart-viewer/ui/ChartViewer.tsx
@@ -1,19 +1,6 @@
-import { useState } from "react";
 import type { DatasetDetail } from "@/shared/lib/backend";
-import type {
-  ChartOptions,
-  ChartType,
-  ComplexViewOption,
-  ScatterMode,
-} from "@/shared/lib/chartTypes";
-import { useChartDataQuery } from "../api/useChartDataQuery";
-import { useDatasetWriteStatusQuery } from "../api/useDatasetWriteStatusQuery";
-import { useFilterTableDataQuery } from "../api/useFilterTableDataQuery";
-import { useCascadeSelection } from "../hooks/useCascadeSelection";
-import {
-  buildChartRequest,
-  deriveChartViewerState,
-} from "../model/chartViewerLogic";
+import { useChartViewerData } from "../hooks/useChartViewerData";
+import { useChartViewerSelection } from "../hooks/useChartViewerSelection";
 import { ChartViewerControls } from "./ChartViewerControls";
 import { ChartWrapper } from "./ChartWrapper";
 import { FilterTable } from "./FilterTable";
@@ -30,99 +17,22 @@ interface ChartViewerProps {
 }
 
 export function ChartViewer({ datasetId, datasetDetail }: ChartViewerProps) {
-  const [chartType, setChartType] = useState<ChartType>("line");
-  const [selectedComplexView, setSelectedComplexView] = useState<
-    ComplexViewOption[]
-  >(["real", "imag"]);
-  const [selectedComplexViewSingle, setSelectedComplexViewSingle] =
-    useState<ComplexViewOption>("mag");
-
-  const [seriesName, setSeriesName] = useState<string | null>(null);
-  const [xColumnName, setXColumnName] = useState<string | null>(null);
-  const [yColumnName, setYColumnName] = useState<string | null>(null);
-
-  const [scatterMode, setScatterMode] = useState<ScatterMode>("complex");
-  const [scatterSeriesName, setScatterSeriesName] = useState<string | null>(
-    null,
-  );
-  const [scatterTraceXName, setScatterTraceXName] = useState<string | null>(
-    null,
-  );
-  const [scatterTraceYName, setScatterTraceYName] = useState<string | null>(
-    null,
-  );
-  const [scatterXName, setScatterXName] = useState<string | null>(null);
-  const [scatterYName, setScatterYName] = useState<string | null>(null);
-  const [scatterBinName, setScatterBinName] = useState<string | null>(null);
-
   const columns = datasetDetail?.columns ?? [];
-  const derived = deriveChartViewerState(columns, {
-    chartType,
-    seriesName,
-    xColumnName,
-    yColumnName,
-    scatterMode,
-    scatterSeriesName,
-    scatterTraceXName,
-    scatterTraceYName,
-    scatterXName,
-    scatterYName,
-    scatterBinName,
-  });
-  const { excludeColumns } = derived;
-
-  const filterTableQuery = useFilterTableDataQuery(
+  const selection = useChartViewerSelection(columns);
+  const { chartData, chartError, filterTableProps } = useChartViewerData({
     datasetId,
-    excludeColumns,
-    Boolean(datasetDetail),
-  );
-  const filterTableData = filterTableQuery.data ?? null;
-
-  const cascade = useCascadeSelection(filterTableData);
-  const filterRow = cascade.resolvedRow ?? null;
-  const hasFilters = (filterTableData?.fields.length ?? 0) > 0;
-  const indexFilters = hasFilters ? filterRow?.valueIndices : undefined;
-
-  useDatasetWriteStatusQuery(datasetId, datasetDetail?.status === "Writing");
-
-  const chartRequest = buildChartRequest({
     datasetDetail,
-    filterTableData,
-    hasFilters,
-    filterRow,
-    selectedComplexView,
-    selectedComplexViewSingle,
-    indexFilters,
-    derived,
+    derived: selection.derived,
+    selectedComplexView: selection.controlState.selectedComplexView,
+    selectedComplexViewSingle: selection.controlState.selectedComplexViewSingle,
   });
-
-  const chartQuery = useChartDataQuery(datasetId, chartRequest);
-  const data: ChartOptions | undefined = chartQuery.data;
-  const chartError = chartQuery.error
-    ? chartQuery.error instanceof Error
-      ? chartQuery.error.message
-      : "Failed to load chart data."
-    : null;
 
   return (
     <div className="flex size-full min-h-0 flex-col overflow-hidden">
       <ChartViewerControls
-        derived={derived}
-        selectedComplexView={selectedComplexView}
-        selectedComplexViewSingle={selectedComplexViewSingle}
-        setChartType={setChartType}
-        setSeriesName={setSeriesName}
-        setXColumnName={setXColumnName}
-        setYColumnName={setYColumnName}
-        setScatterMode={setScatterMode}
-        setScatterSeriesName={setScatterSeriesName}
-        setScatterTraceXName={setScatterTraceXName}
-        setScatterTraceYName={setScatterTraceYName}
-        setScatterXName={setScatterXName}
-        setScatterYName={setScatterYName}
-        setScatterBinName={setScatterBinName}
-        setSelectedComplexView={setSelectedComplexView}
-        setSelectedComplexViewSingle={setSelectedComplexViewSingle}
+        derived={selection.derived}
+        controlState={selection.controlState}
+        actions={selection.actions}
       />
 
       {chartError ? (
@@ -138,28 +48,13 @@ export function ChartViewer({ datasetId, datasetDetail }: ChartViewerProps) {
         <ResizablePanelGroup orientation="vertical" className="h-full min-h-0">
           <ResizablePanel defaultSize={70} minSize={35} className="min-h-0">
             <div className="h-full min-h-0">
-              <ChartWrapper data={data} />
+              <ChartWrapper data={chartData} />
             </div>
           </ResizablePanel>
           <ResizableHandle withHandle />
           <ResizablePanel defaultSize={30} minSize={20} className="min-h-0">
             <div className="h-full min-h-0">
-              <FilterTable
-                data={filterTableData ?? undefined}
-                mode={cascade.state.mode}
-                onModeChange={cascade.setMode}
-                selectedRowIndex={filterRow?.index ?? null}
-                onSelectRow={cascade.selectRow}
-                selectedValueIndices={cascade.selectedValueIndices}
-                onSelectFieldValue={(fieldIndex, valueIndex) => {
-                  if (!filterTableData) return;
-                  cascade.selectFieldValue(
-                    fieldIndex,
-                    valueIndex,
-                    filterRow?.index ?? null,
-                  );
-                }}
-              />
+              <FilterTable {...filterTableProps} />
             </div>
           </ResizablePanel>
         </ResizablePanelGroup>
