+++ crates/fricon-ui/frontend/src/features/dataset-table/api/useDatasetTableQuery.ts
@@ -1,0 +1,61 @@
+import {
+  keepPreviousData,
+  useQuery,
+  useQueryClient,
+} from "@tanstack/react-query";
+import { listDatasets } from "@/shared/lib/backend";
+import {
+  buildDatasetListOptions,
+  deriveHasMore,
+  type DatasetQueryKey,
+  type DatasetQueryParams,
+} from "./datasetTableShared";
+
+export function useDatasetTableQuery(
+  queryParams: DatasetQueryParams,
+  visibleCount: number,
+  incrementVisibleCount: () => Promise<void>,
+) {
+  const queryClient = useQueryClient();
+  const datasetQueryKey: DatasetQueryKey = [
+    "datasets",
+    "list",
+    queryParams,
+    visibleCount,
+  ];
+
+  const datasetsQuery = useQuery({
+    queryKey: datasetQueryKey,
+    queryFn: () =>
+      listDatasets(
+        buildDatasetListOptions(queryParams, {
+          limit: visibleCount,
+          offset: 0,
+        }),
+      ),
+    placeholderData: keepPreviousData,
+  });
+
+  const datasets = datasetsQuery.data ?? [];
+
+  const refreshDatasets = async () => {
+    await queryClient.invalidateQueries({ queryKey: datasetQueryKey });
+  };
+
+  const hasMore = datasetsQuery.isPlaceholderData
+    ? true
+    : deriveHasMore(datasets.length, visibleCount);
+
+  const loadNextPage = () => {
+    if (datasetsQuery.isFetching || !hasMore) return Promise.resolve();
+    return incrementVisibleCount();
+  };
+
+  return {
+    datasetQueryKey,
+    datasets,
+    hasMore,
+    refreshDatasets,
+    loadNextPage,
+  };
+}
